refactor: refactored components in its own files and folders
