+++ src/Components/ProposalCard/ExpandedActionsList/index.ts
@@ -1,0 +1,1 @@
+export { default as ExpandedActionsList } from './ExpandedActionsList';
