+++ src/Components/ProposalCard/ProposalCardWinningOption/index.ts
@@ -1,0 +1,1 @@
+export { default as ProposalCardWinningOption } from './ProposalCardWinningOption';
